Refactored ParseFlags out of CSVReader
